fix: api middleware ignore unknown paths
