--- indexer/pkg/api/middleware/path_normalizer.go
@@ -1,16 +1,0 @@
-package middleware
-
-import (
-	"strings"
-)
-
-// RemoveMessageIDFromPath normalizes verifierresults paths by replacing message IDs with a placeholder.
-// Always returns true to track all endpoints in the indexer.
-func RemoveMessageIDFromPath(path string) (string, bool) {
-	if strings.Contains(path, "/verifierresults/") {
-		// Normalize to a canonical path with a placeholder for the message ID.
-		return "/verifierresults/:messageID", true
-	}
-
-	return path, true
-}
